proj-titanic-prediction/pyproject.toml
[tool.poetry]
name = "proj-titanic-prediction"
version = "0.1.0"
description = ""
authors = ["apicharto <155867145+apicharto@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.27"
pymysql = "^1.1.0"
pandas = "^2.2.1"
streamlit = "^1.31.1"
google-cloud-bigquery = "^3.17.2"
db-dtypes = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
